+++ tests/testthat/test_desc.R
@@ -10,7 +10,7 @@ vars <- c("elevation", "slope", "aspect", "land_cover_type",
           "texture1", "texture2", "texture3", "soilvolextract", "soilvolsat", "soilvolfc", "soilvolwp", "soilvolcurr",
           "basal_area", "leaf_area_index", "foliar_biomass", "fuel")
 vars_ws <- c(vars,
-             "num_neigh", "water_order", "outlets", "channel","depth_to_bedrock","bedrock_porosity", "bedrock_conductivity",
+             "channel","depth_to_bedrock","bedrock_porosity", "bedrock_conductivity",
              "aquifer_elevation", "depth_to_aquifer","aquifer_volume", "snowpack")
 test_that("Can extract variables from 'sf' object",{
   expect_s3_class(extract_variables(ypts, vars = vars, SpParams = SpParamsMED), "sf")